+++ lib/cypress_rails/configuration.rb
@@ -1,18 +1,10 @@
 module CypressRails
-  class << self
-    attr_accessor :configuration
-  end
-
-  def self.configure
-    self.configuration ||= Configuration.new
-    yield(configuration)
-  end
-
   class Configuration
-    attr_accessor :before_each
+    attr_accessor :before_each, :scripts_path
 
     def initialize
       @before_each = -> {}
+      @scripts_path = "spec/cypress/seeds"
     end
   end
 end